Bug 1838120 - Unify Fenix buildSrc/Config with the ConfigPlugin
