+++ fenix/app/build.gradle
@@ -289,7 +289,9 @@ android.applicationVariants.all { variant ->
 
             println("versionCode for $abi = $versionCodeOverride, isMozillaOnline = $isMozillaOnline")
 
-            output.versionNameOverride = versionName
+            if (versionName != null) {
+                output.versionNameOverride = versionName
+            }
             output.versionCodeOverride = versionCodeOverride
         }
     } else if (gradle.hasProperty("localProperties.branchBuild.fenix.version")) {
